+++ ui/templates/delete_user.html
@@ -1,0 +1,20 @@
+{{define "content"}}
+
+<div class="container">
+    <h2 class="tls-title">Delete {{.Name}}?</h2>
+    <div class="tls-form">
+        
+        <form enctype="multipart/form-data" class="form-horizontal needs-validation" novalidate
+            action="/ui/user/id/{{.Name}}/delete" method="POST" novalidate>
+            <input id="id" id="name" value="{{.Name}}" hidden />
+            {{.CSRFField}}
+
+            <a href="/ui/user/id/{{.Name}}/" class="btn btn-outline-primary" type="button">Cancel</a>
+            <button class="btn btn-danger" type="submit" id="submit-form">Delete</button>
+            <button style="display:none;" id="form-working-message" class="btn btn-danger" disabled><i
+                    class="fas fa-spinner glyphicon-spin"></i> Loading...</button>
+        </form>
+
+    </div>
+</div>
+{{end}}
