+++ ui/templates/layout.html
@@ -14,6 +14,12 @@
                 <a class="nav-item nav-link" href="/ui/dashboard">Dashboard</a>
             </div>
             <div class="navbar-nav">
+                <a class="nav-item nav-link" href="/ui/certificates">Certificates</a>
+            </div>
+            <div class="navbar-nav">
+                <a class="nav-item nav-link" href="/ui/users">Users</a>
+            </div>
+            <div class="navbar-nav">
                 <a class="nav-item nav-link" href="/ui/certificate/create">Create</a>
             </div>
