+++ ui/templates/users.html
@@ -1,0 +1,103 @@
+{{define "content"}}
+<h2 class="tls-title">Users <a href="/ui/user/create">+</a></h2>
+
+<div class="tls-page">
+    <div class="row">
+        <div class="col-12 order-md-1">
+            <table class="table table-hover">
+                <thead>
+                    <tr>
+                        <th scope="th" style="width: 5rem;"></th>
+                        <th scope="col">Name</th>
+                        <th scope="col" class="text-right">Role</th>
+                    </tr>
+                </thead>
+                <tbody>
+                    {{ range .Users }}
+                    <tr id="user-row-{{.Name}}">
+                        <td class="text-center" style="width:5rem">
+                            <a data-toggle="tooltip" data-placement="bottom" class="table-icon text-primary"
+                                href="/ui/user/id/{{.Name}}/edit" title="Edit user">
+                                <i class="fas fa-pencil-alt"></i>
+                            </a>
+                            <a data-toggle="tooltip" data-placement="bottom" class="table-icon text-danger"
+                                title="Delete user" href="javascript:userDeleteModal('{{.Name}}', '{{.Name}}');">
+                                <i class="fas fa-trash"></i>
+                                </a>
+                        </td>
+                        <td>
+                            <a data-toggle="tooltip" data-placement="bottom" href="/ui/user/id/{{.Name}}"
+                                title="View user">{{.Name}}</a>
+                        </td>
+                        <td class="text-right">
+                            {{.Role}}
+                        </td>
+                    </tr>
+                    {{end}}
+                </tbody>
+            </table>
+        </div>
+    </div>
+</div>
+
+<div id="tls-delete-modal" class="modal fade" tabindex="-1" role="dialog" aria-labelledby="deleteConfirmModal" aria-hidden="true">
+    <div class="modal-dialog">
+        <div class="modal-content">
+            <div class="modal-header bg-danger">
+                <h5 class="modal-title text-light">Warning!</h5>
+            </div>
+            <div class="modal-body">
+                <p>Deleting <span class="font-weight-bold" id="user-delete-name"></span> cannot be undone.</p>
+                <p class="font-weight-bold">Are you sure you want to delete this user?</p>
+            </div>
+            <div class="modal-footer">
+                <button data-dismiss="modal" id="Cancel" type="button" class="btn btn-outline-secondary">Cancel</button>
+                <a role="button" id="delete-btn" href="#" class="btn btn-danger pull-right">Delete</a>
+                <button style="display:none;" id="form-working-message" class="btn btn-primary" disabled><i
+                    class="fas fa-spinner glyphicon-spin"></i> Loading...</button>
+            </div>
+        </div><!-- /.modal-content -->
+    </div>
+</div><!-- /.modal -->
+<script>
+
+    window.onload = function () {
+        $.ajaxPrefilter(function (options, originalOptions, jqXHR) {
+            jqXHR.setRequestHeader('X-CSRF-Token', $("input[name=csrf_token]")[0].value);
+        });
+    }
+
+    function userDeleteModal(id, name) {
+        $("#user-delete-name").html(name);
+        $("#delete-btn").attr("href", "javascript:userDeleteConfirmed('" + id + "')");
+        $("#tls-delete-modal").modal('show');
+    }
+
+    function userDeleteConfirmed(id) {
+        var url = "/ui/user/id/" + id + "/delete";
+        $("#delete-btn").hide();
+        $("#form-working-message").show();
+        $.ajax(url,{
+            url: url,
+            type: 'POST',
+            dataType: 'html',
+            // data: { token: id }, //could send token here if the other funtion doen't work
+            contentType: "application/x-www-form-urlencoded",
+            success: function (result) {
+                token = result;
+                $("#user-row-" + id).remove();
+                $("#tls-delete-modal").modal('hide');
+                $("#form-working-message").hide();
+                $("#delete-btn").show()
+            },
+            error: function (passParams,status, err) {
+                console.log("Error in userDeleteConfirmed ajax call.");
+                console.log(passParams,status,err);
+                $("#form-working-message").hide();
+                $("#delete-btn").show()
+            }
+        });
+
+    }
+</script>
+{{end}}
