+++ ui/templates/view_certificate.html
@@ -1,24 +1,5 @@
 {{define "content"}}
 
-<!-- 
-    <div class="col-md-8 order-md-1">
-      <h4 class="mb-3">Certificate</h4>
-
-        <div class="row">
-          <div class="col-md-6 mb-3">
-            <label for="id">ID</label>
-            <input type="text" class="form-control" id="id" placeholder="" value="{{.Cert.ID}}" readonly>
-          </div>
-          <div class="col-md-6 mb-3">
-            <label for="secret">Secret</label>
-            <div class="input-group">
-              <input type="text" class="form-control" id="secret" placeholder="" value="{{.Cert.Secret}}" readonly>
-            </div>
-          </div>
-        </div>
-
-    </div> -->
-
 
 
 <h2 class="tls-title">{{.Cert.CommonName}}</h2>
